+++ core/CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file. This projec
 
 ### Changed
 
++ [#1827](https://github.com/luyadev/luya/issues/1827) ResponseCache using PageCache filter and mark $actionsCallable and $actions as deprecated.
 + [#1821](https://github.com/luyadev/luya/issues/1821) Remove public properties from `luya\traits\CacheableTrait` in order to prevent conflicts, as they only contain fallback informations.
 + [#1820](https://github.com/luyadev/luya/issues/1820) Cover sensitive data in robots filter post data.